requirements.txt
elevenlabs==0.2.27
python-dotenv==1.0.0
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
